tests/functional/requirements.txt
pytest==7.2.1
pytest-asyncio==0.18.0
aiohttp==3.8.3
elasticsearch==8.6.0
python-dotenv==0.21.0
pydantic==1.10.3
redis==4.1.4
aioredis==2.0.1
